refactor(android): removed redundant service layer
- service class
- service class tests
